requirements.txt
aiobotocore==2.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
astropy==7.1.0
astropy-iers-data==0.2025.8.25.0.36.58
astroquery==0.4.10
attrs==25.3.0
autograd==1.8.0
beautifulsoup4==4.13.5
bokeh==3.8.0
botocore==1.39.11
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fbpca==1.0
filelock==3.13.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
html5lib==1.1
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
keyring==25.6.0
kiwisolver==1.4.9
lightkurve==2.5.1
MarkupSafe==3.0.2
matplotlib==3.10.5
memoization==0.4.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
narwhals==2.2.0
networkx==3.3
numpy==2.3.2
oktopus==0.1.2
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
propcache==0.3.2
pyerfa==2.0.1.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyvo==1.7
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.5
s3fs==2025.7.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==70.2.0
six==1.17.0
soupsieve==2.8
streamlit==1.50.0
sympy==1.13.3
tensorboard==2.20.0
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.5.0
webencodings==0.5.1
wrapt==1.17.3
xyzservices==2025.4.0
yarl==1.20.1
